DTP-78 Not working unit test for adapter because project cant reference the lib that holds the treeviewitem
